handle exception user and role

+++ user-service/src/main/java/hcmus/alumni/userservice/controller/RoleServiceController.java
@@ -5,6 +5,8 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.InvalidDataAccessApiUsageException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -12,10 +14,8 @@ import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.web.bind.MissingServletRequestParameterException;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -24,6 +24,7 @@ import hcmus.alumni.userservice.dto.IPermissionDto;
 import hcmus.alumni.userservice.dto.IRoleDto;
 import hcmus.alumni.userservice.dto.IRoleWithoutPermissionsDto;
 import hcmus.alumni.userservice.dto.RoleRequestDto;
+import hcmus.alumni.userservice.exception.AppException;
 import hcmus.alumni.userservice.model.PermissionModel;
 import hcmus.alumni.userservice.model.RoleModel;
 import hcmus.alumni.userservice.repository.PermissionRepository;
@@ -43,13 +44,6 @@ public class RoleServiceController {
     @Autowired
     private PermissionRepository permissionRepository;
 
-    @ExceptionHandler(MissingServletRequestParameterException.class)
-    public void handleMissingParams(MissingServletRequestParameterException ex) {
-        String name = ex.getParameterName();
-        System.out.println(name + " parameter is missing");
-        // Actual exception handling
-    }
-
     @GetMapping("/permissions")
     public ResponseEntity<HashMap<String, Object>> getAllPermissions() {
         List<IPermissionDto> permissions = permissionRepository.findAllPermissions();
@@ -67,11 +61,17 @@ public class RoleServiceController {
             @RequestParam(value = "order", required = false, defaultValue = "asc") String order) {
         HashMap<String, Object> result = new HashMap<String, Object>();
 
-        Pageable pageable = PageRequest.of(page, pageSize, Sort.by(Sort.Direction.fromString(order), orderBy));
-        Page<IRoleWithoutPermissionsDto> roles = roleRepository.searchRoles(name, pageable);
+        try {
+            Pageable pageable = PageRequest.of(page, pageSize, Sort.by(Sort.Direction.fromString(order), orderBy));
+            Page<IRoleWithoutPermissionsDto> roles = roleRepository.searchRoles(name, pageable);
+            result.put("totalPages", roles.getTotalPages());
+            result.put("roles", roles.getContent());
+        } catch (IllegalArgumentException e) {
+            throw new AppException(80100, "Tham số order phải là 'asc' hoặc 'desc'", HttpStatus.BAD_REQUEST);
+        } catch (InvalidDataAccessApiUsageException e) {
+            throw new AppException(80101, "Tham số orderBy không hợp lệ", HttpStatus.BAD_REQUEST);
+        }
 
-        result.put("totalPages", roles.getTotalPages());
-        result.put("roles", roles.getContent());
         return ResponseEntity.status(HttpStatus.OK).body(result);
     }
 
@@ -79,7 +79,7 @@ public class RoleServiceController {
     public ResponseEntity<IRoleDto> getRole(@PathVariable Integer id) {
         Optional<IRoleDto> role = roleRepository.findRoleById(id);
         if (role.isEmpty()) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+            throw new AppException(80200, "Không tìm thấy vai trò", HttpStatus.NOT_FOUND);
         }
         return ResponseEntity.status(HttpStatus.OK).body(role.get());
     }
@@ -88,7 +88,11 @@ public class RoleServiceController {
     @PostMapping("")
     public ResponseEntity<String> postRole(@RequestBody RoleRequestDto requestingRole) {
         RoleModel role = new RoleModel(requestingRole);
-        roleRepository.save(role);
+        try {
+            roleRepository.save(role);
+        } catch (DataIntegrityViolationException e) {
+            throw new AppException(80300, "Không tìm thấy quyền", HttpStatus.NOT_FOUND);
+        }
         return ResponseEntity.status(HttpStatus.OK).body(null);
     }
 
@@ -97,7 +101,7 @@ public class RoleServiceController {
     public ResponseEntity<String> putRole(@PathVariable Integer id, @RequestBody RoleRequestDto requestingRole) {
         Optional<RoleModel> roleOptional = roleRepository.findById(id);
         if (roleOptional.isEmpty()) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+            throw new AppException(80400, "Không tìm thấy vai trò", HttpStatus.NOT_FOUND);
         }
 
         RoleModel role = roleOptional.get();
@@ -119,15 +123,24 @@ public class RoleServiceController {
             isPut = true;
         }
 
-        if (isPut) {
-            roleRepository.save(role);
+        try {
+            if (isPut) {
+                roleRepository.save(role);
+            }
+        } catch (DataIntegrityViolationException e) {
+            throw new AppException(80401, "Không tìm thấy quyền", HttpStatus.NOT_FOUND);
         }
+
         return ResponseEntity.status(HttpStatus.OK).body(null);
     }
 
     @PreAuthorize("hasAuthority('User.Role.Delete')")
     @DeleteMapping("/{id}")
     public ResponseEntity<String> deleteRole(@PathVariable Integer id) {
+        // Default roles, cannot delete
+        if (id >= 1 && id <= 5) {
+            throw new AppException(80500, "Không thể xóa vai trò mặc định", HttpStatus.BAD_REQUEST);
+        }
         roleRepository.deleteById(id);
         return ResponseEntity.status(HttpStatus.OK).body(null);
     }
